solana: fix merge conflict; reorganize
