+++ solana/programs/swap-layer/src/processor/complete/transfer/mod.rs
@@ -1,8 +1,8 @@
-mod relay;
-pub use relay::*;
-
 mod direct;
 pub use direct::*;
 
 mod payload;
 pub use payload::*;
+
+mod relay;
+pub use relay::*;
